Refactor TransformAction.
Make TransformAction just responsible for figuring out what the
moment-by-moment transformation matrices should be, and factor out a
separate class TransformLayerSink that is responsible for actually
creating a Transform event and adding to it. This makes it easier for
future changes to add different possible sinks for the same
TransformAction. For example, we can use this to transform individual
strokes instead of entire layers.