Fix seq ids of cached objects.
This just makes more sense, and also fixes a bug where rectangles that
are drawn at a fixed instant in time (and viewed at that same instant)
will be inconsistently drawn when they are transformed. Sometimes only a
few of the stroke events are drawn instead of all of them.