Fix visibility display bug in timeline.
There is logic to colour a layer's timeline based on whether or not that
layer is currently visible. This logic includes reasoning about events
that happen before the set of events currently in the timeline viewport,
since those events might affect visibility within the current viewport.
Because of an off-by-one error, we would never consider events in the
first bucket when doing this logic.