Fix body size bug.
Since layer canvases have relative position, they are removed from
normal layout flow and don't contribute to <body> size. This means that
when we install mouse event handlers on <body>, those events don't
necessarily fire near the bottom of the canvas. We fix this by manually
specifying the height that the layer_set container should be, so that
<body> covers the entire page.