Mix audio on client side and upload to server.
On export, we now mix audio on client side using an OfflineAudioContext
and then upload that audio to the server.

This is a change from previous plan to mix on server side. Client-side
will be better because it's exactly the same code to mix for interactive
playback and to mix for final product, so there are no surprises during
export.

Still need to combine audio and video on server side.