Use coalesced mouse events for smoother paint.
Take advantage of the PointerEvent.getCoalescedEvents() method to access
higher frequency mouse events for the PaintAction so that lines are
smoother.

There is an option to disable this, because in the rare case of drawing
over a scene that is transforming heavily, it seems to cause more jagged
lines. Haven't really investigated why.