Fix missing image bug.
Stencil and layer background images load asynchronously. Before this
change, if we try to draw them before they are fully loaded, we may end
up in a state where we think our drawing is done but it is actually
incomplete. After this change, we keep track of whether images are fully
loaded or not. If there are any missing images, we don't consider the
drawing to be complete, and we try again later.